+++ include/Queue.h
@@ -6,14 +6,14 @@
 
 typedef struct { // Circular Queue
     genVec* arr;
-    u32 head;   // dequeue (head + 1) % capacity
-    u32 tail;   // enqueue (head + size) % capacity
-    u32 size;     
+    u64 head;   // dequeue (head + 1) % capacity
+    u64 tail;   // enqueue (head + size) % capacity
+    u64 size;     
 } Queue;
 
 
-Queue* queue_create(u32 n, u16 data_size, genVec_copy_fn copy_fn, genVec_move_fn move_fn, genVec_delete_fn del_fn);
-Queue* queue_create_val(u32 n, const u8* val, u16 data_size, genVec_copy_fn copy_fn, genVec_move_fn move_fn, genVec_delete_fn del_fn);
+Queue* queue_create(u64 n, u32 data_size, genVec_copy_fn copy_fn, genVec_move_fn move_fn, genVec_delete_fn del_fn);
+Queue* queue_create_val(u64 n, const u8* val, u32 data_size, genVec_copy_fn copy_fn, genVec_move_fn move_fn, genVec_delete_fn del_fn);
 void queue_destroy(Queue* q);
 void queue_clear(Queue* q);
 void queue_reset(Queue* q);
@@ -26,7 +26,7 @@ void queue_peek(Queue* q, u8* peek);
 const u8* queue_peek_ptr(Queue* q);
 
 
-static inline u32 queue_size(Queue* q) {
+static inline u64 queue_size(Queue* q) {
     CHECK_FATAL(!q, "queue is null");
     return q->size;
 }
@@ -36,7 +36,7 @@ static inline u8 queue_empty(Queue* q) {
     return q->size == 0;
 }
 
-static inline u32 queue_capacity(Queue* q) {
+static inline u64 queue_capacity(Queue* q) {
     CHECK_FATAL(!q, "queue is null");
     CHECK_FATAL(!q->arr, "queue->arr is null");
     return genVec_capacity(q->arr);
